5단계 비밀번호 조건 수정

+++ src/main/java/com/sparta/springpersonalboard/service/BoardService.java
@@ -53,9 +53,15 @@ public class BoardService {
         return id;
     }
 
-    public Long deleteBaord(Long id) {
+    public Long deleteBaord(Long id, BoardRequestDto requestDto) {
         Board board = findBoard(id);
-        boardRepository.delete(board);
+        String password = requestDto.getPassword();
+
+        if (password.equals(board.getPassword())) {
+            boardRepository.delete(board);
+        } else {
+            System.out.println("비밀번호가 일치하지 않습니다.");
+        }
         return id;
     }
